# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Structural AI API",
    "description": "Deep Chomskyan syntactic analysis API for EdTech and research applications. Provides automated linguistic annotation including noun phrase structures, verb phrase analysis, morphemic breakdowns, and structural ambiguity detection.",
    "version": "1.0.0",
    "contact": {
      "email": "support@structuralai.api"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.structuralai.app/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/analyze/text": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze text for syntactic structure",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Text to analyze (max 50,000 tokens)",
                    "example": "The quick brown fox jumps over the lazy dog."
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "includeMorphemes": {
                        "type": "boolean",
                        "default": true
                      },
                      "detectAmbiguities": {
                        "type": "boolean",
                        "default": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or text too long"
          },
          "401": {
            "description": "Authentication required"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/analyze/audio": {
      "post": {
        "tags": ["Analysis"],
        "summary": "Analyze audio file (transcribe + syntactic analysis)",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["audio"],
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file (wav, mp3, ogg, webm, flac, m4a). Max 50MB."
                  },
                  "language": {
                    "type": "string",
                    "description": "Source language code (auto-detect if not specified)",
                    "example": "en"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid audio format"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Your API key (also accepts Bearer token in Authorization header)"
      }
    },
    "schemas": {
...
      "TextAnalysisResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "status": {"type": "string", "enum": ["completed", "failed"]},
          "processingTimeMs": {"type": "number"},
          "result": {
            "type": "object",
            "properties": {
              "input": {"type": "object"},
              "structural": {"type": "object"},
              "syntactic": {"type": "object"},
              "metadata": {"type": "object"}
            }
          },
          "billing": {
            "type": "object",
            "properties": {
              "tokens": {"type": "number"},
              "billable": {"type": "boolean"}
            }
          }
        }
      },
      "AudioAnalysisResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "status": {"type": "string"},
          "processingTimeMs": {"type": "number"},
          "result": {"type": "object"},
          "billing": {
            "type": "object",
            "properties": {
              "audioSeconds": {"type": "number"},
              "audioMinutes": {"type": "number"},
              "billable": {"type": "boolean"}
            }
          }
        }
      },
...
    }
  }
}